Don't compile nondigest assets in separate process
Before this patch rake task for compiling assets was running separate
process to compile nondigest assets, instead of just invoking the task
in the same process. In order to make it work without having 2 processes
we need to clear sprockets cache, which is impossible with
`Sprockets::Index` - it's immutable. That's why I'm making a copy of
`Environment` before converting it to index, to be able to clear the
cache and use it for nondigest version. It's a bit hacky, but since
`Index` just references most of the things in `Environment` it should
work just fine.