Detect asset changes by adding source_digests to manifest.yml
This allows us to quickly determine if the asset needs to be
recompiled. If nothing has changed in the source assets,
then compilation is skipped.
Also, instead of compiling twice for non-digest assets,
digest assets are copied instead, and the known digests are
stripped from the body of css and js files.
The copied js/css files are also gzipped.