Undefine Rails' assets method before redefining it, to avoid warnings
Rails test suite keeps raising:

    warning: method redefined; discarding old assets
    warning: method redefined; discarding old assets=

Because sprockets-rails is monkey-patching Rails::Application, redefining
the assets method, and adding an attr_writer for it.

This undefines the existing method before adding the new one, avoiding
the warning. The writer call is also removed, since Rails defines an
attr_accessor that already adds it.