Allow setting keep explicitly
The number of backup copies of an asset that are kept after running `assets:clean` is currently hardcoded at 2. You currently cannot configure this number. 

Instead of pushing the ability to change this via an ENV var sstephenson/sprockets#461 it was suggested that this ability be done at the rake level which results in a clean API.

So if you do not want to keep any copies of assets you can run

```sh
$ rake assets:clean[0]
```

You can keep 1 copy of assets by running:

```sh
$ rake assets:clean[1]
```

If no value is provided 2 copies will be stored.

ATP