Instead of complaining, add the dependency automatically
We need to use *both* `depend_on` and `depend_on_asset`, because the
latter (bugfully?) considers changes to the referenced asset's
dependencies, but not the file itself (making it entirely useless for an
image).

In the corresponding test, we call a protected Sprockets method... by
design, this isn't something we're supposed to care about / access. But
actually manipulating assets to ensure the dependency has taken hold
seems rather like overkill.