If debug is true, async options should be ignored.
The async attribute causes script tags to be executed as soon as they are
downloaded. Often, developers will use the order of the scripts inside of their
manifest file as a kind of implicit dependency ordering - Jquery has to execute
before jquery_ujs, and so on.

When debug is on, async will cause these scripts to be executed in an unreliable
order, breaking the application's scripts. When debug is on, we should simply
omit the async tag.

We do not need to omit defer, as browsers *should* still execute deferred scripts
in the order they appear in the document.