[Close #321] Clear Asset precompile cache on request
When in debug mode we want to see if any new files have been added to our precompile list by nature of them being added to disk. For example if you add an image  to `images` and you have linked the directory:

```
//= link_directory ../images
```

This change also eliminates the need to restart the server in development when changes are made when using a manifest.js and Sprockets 4.