Allow using a custom assets configuration manifest location
At the moment the only allowed path is the default one but
there could be the need to set a custom configuration manifest
location.

This commit introduces a new setting `config.assets.config_manifest`
that allows to change the location of this manifest.
This is an example configuration:

```
config.assets.config_manifest = File.expand_path('my_custom_path/manifest.js', __dir__)
```