Warn instead of raising on missing manifest.js
With a pristine (`rails new`) Rails 8.0 app, the `app/assets/config/manifest.js` file
is no longer automatically generated. For engines still using sprockets,
that is not a huge issue, because they have and add their own manifest
files for their assets to the list of precompiled assets.

However, scripts like the Solidus bin/sandbox script[1] first run `rails
new`, then add the needed engines to the sandbox's `Gemfile`, then run
`rails db:migrate` and their respective install generators. With Rails
8, these scripts now fail before anything is possible in Ruby, because
the initializer of `Sprockets::Railtie` will raise an error on startup.

This commit changes the behavior of the Railtie to issue a deprecation
warning instead of raising an Exception.

[1] https://github.com/solidusio/solidus/blob/main/bin/sandbox#L44-L99